Add files via upload
Technologies Used:

HTML: Markup language for structuring the content and layout of the web pages.
CSS: Stylesheet language for styling the visual presentation of the application.
JavaScript: Programming language for dynamic behavior, user interactions, and API integration.
Unsplash API: External API used to retrieve high-quality images based on user queries.
How to Use:

Open the Image Search web application in a web browser.
Enter a search query in the provided input field.
Press the "Search" button to fetch and display relevant images.
Scroll through the results, and click on images for more details on the Unsplash website.
Use the "Load More" button to fetch additional images dynamically.
Project Outcome:

The Image Search project provides users with a seamless and visually engaging platform to explore and discover a wide array of images. By combining the power of HTML, CSS, JavaScript, and the Unsplash API, the project delivers a responsive and feature-rich image search experience, catering to users' preferences and interests.